Extract ExecutionContext to executor/context.rs module
Second small step in executor.rs refactoring. This change moves only
the ExecutionContext type to a dedicated module while maintaining
all existing APIs and functionality.

Continues the systematic approach of small, focused extractions
to break down the large executor.rs file into manageable modules.
All tests pass with no behavioral changes.
