+++ PLANNING.md
@@ -54,7 +54,7 @@ EventCore has successfully completed all initially planned phases (1-20), includ
   - [x] `executor/retry.rs` - Retry and circuit breaker logic (RetryConfig, RetryPolicy extracted)
   - [ ] `executor/stream_discovery.rs` - Stream discovery iteration logic
   - [ ] `executor/validation.rs` - Command validation
-  - [ ] `executor/context.rs` - Execution context management
+  - [x] `executor/context.rs` - Execution context management (ExecutionContext extracted)
 
 #### 2. Refactor cqrs/rebuild.rs::rebuild function (189 lines) - **HIGH**
 **Problem**: Complex rebuild logic with multiple responsibilities
